fix: MyPy batch 5 (100 errors from 19 files)
Fixed 100 MyPy type checking errors across the following files:

victor/optimizations/algorithms.py:
- Fixed Lazy class to be Generic[T] for proper TypeVar usage
- Fixed lru_cache decorator type signatures
- Added type annotations for cache variables
- Fixed ItemsView type parameter
- Added type ignore for _timed_caches iteration

victor/optimizations/concurrency.py:
- Added type parameters to Queue (Queue[float], Queue[T])
- Added return type annotation to __init__
- Fixed os.cpu_count() None handling with fallback to 1

victor/optimizations/database.py:
- Fixed cached_query decorator to use Awaitable[T] instead of T
- Added Awaitable import
- Renamed variables to avoid redefinition

victor/optimizations/lazy_loader.py:
- Added TypeVar F for decorator functions
- Fixed lazy_load return type with cast
- Reorganized descriptor access to avoid unreachable code error
- Added cast import

victor/optimizations/memory.py:
- Changed get_stats() return type from Dict[str, int] to Dict[str, float]
- Fixed pool_stats type from Dict[str, int] to Dict[str, float]
- Fixed generation type from int to str for dict compatibility

victor/optimizations/network.py:
- Added type parameters to Future and Task (Future[Any])
- Fixed base_url None handling with default value
- Added type annotation for response_data with cast
- Added assert for client None check

victor/optimizations/parallel_executor.py:
- Added type parameters to TaskInput (TaskInput[Any])
- Added type annotations for coro variables
- Fixed Queue type parameters (Queue[Tuple[int, TaskInput[Any]]])
- Fixed TaskWithPriority argument handling
- Removed unused type: ignore comment
- Added cast for batch_results_with_exceptions

victor/processing/completion/manager.py:
- Fixed stream_coro to await before async iteration

victor/rag/advanced_rag.py:
- Fixed DocumentChunk.source to use metadata.get("source", "")
- Fixed sources_list type annotation

victor/rag/assistant.py:
- Removed StageDefinition import (not exported)
- Changed name, description, version to explicit type annotations

victor/rag/capabilities.py:
- Added cast(Dict[str, Any]) for get_capability_config results
- Added return type annotation to _get_provider
- Added cast import

victor/rag/chunker.py:
- Added type annotations for current_chunk and current_batch lists

victor/rag/demo_docs.py:
- Added return type annotation to main()
- Fixed set type parameter (set[str])

victor/rag/demo_sec_filings.py:
- Added type annotations for filings list
- Changed function parameters to Optional[str]
- Added return type annotation to main() and list_companies()
- Added type: ignore for filter_symbol argument

victor/rag/document_store.py:
- Added assert for _db and _table None checks
- Fixed do_hybrid_search and do_vector_search return types
- Added cast for embedding.tolist() result
- Added cast import

victor/rag/enrichment.py:
- Changed enhancement_techniques to Optional[List[str]]

victor/rag/escape_hatches.py:
- Added cast for sorting key function
- Fixed text None handling in len() call
- Added cast import

victor/rag/handlers.py:
- Added _exec_ctx parameter to tool_registry.execute
- Fixed _chunk_document argument types with int() conversions

victor/rag/mode_config.py:
- Added return type annotation to __init__
- Added ValueError for missing mode in get_rag_mode_config

+++ victor/optimizations/memory.py
@@ -59,7 +59,7 @@ class MemoryStats:
     total_objects: int = 0
     total_size: int = 0
     gc_counts: tuple[int, int, int] = field(default_factory=lambda: (0, 0, 0))
-    pool_stats: Dict[str, Dict[str, int]] = field(default_factory=dict)
+    pool_stats: Dict[str, Dict[str, float]] = field(default_factory=dict)
 
     def to_dict(self) -> Dict[str, Any]:
         """Convert to dictionary."""
@@ -405,7 +405,7 @@ class MemoryOptimizer:
         stats.gc_counts = gc.get_count()
 
         # Get pool stats
-        stats.pool_stats = {name: pool.get_stats() for name, pool in self._pools.items()}  # type: ignore[assignment]
+        stats.pool_stats = {name: pool.get_stats() for name, pool in self._pools.items()}
 
         return stats
